Removed duplicate files. Refactored some prints to be more readable. Added FIXME: on BogoSort.
